+++ backend/common.go
@@ -1,32 +1,23 @@
 package backend
 
-import (
-	"database/sql"
-	"time"
-	"log"
-	"errors"
+import "database/sql"
+
+const (
+	CONFLICT = iota
+	DUPLICATE = iota
+	MISSING = iota
 )
 
-func StoreLog(db *sql.DB, uid int, message string) {
-	_, err := db.Exec("insert into logs(date, uid, message) values($1,$2,$3)", time.Now(), uid, message)
-	if (err != nil) {
-		log.Printf("Unable to store a log entry: %s", err)
-	}
-	log.Printf("%d - %s\n", uid, message)
+type WintsError struct {
+	Kind int
+	Err error
 }
 
-
-type BackendError struct {
-	Status int
-	message string
-
+func (w *WintsError) Error() string {
+	return w.Err.Error()
 }
 
-func (err *BackendError) Error() string {
-	return err.message
-}
-
-func SingleUpdate(db *sql.DB, q string, args ...interface{}) error {
+func SingleUpdate(db *sql.DB, errNoUpdate error, q string, args ...interface{}) error {
 	res, err := db.Exec(q, args...)
 	if err != nil {
 		return err
@@ -36,7 +27,7 @@ func SingleUpdate(db *sql.DB, q string, args ...interface{}) error {
 		return err
 	}
 	if nb != 1 {
-		return errors.New("No updates")
+		return errNoUpdate
 	}
 	return nil
 }