+++ backend/defenses.go
@@ -1,6 +1,14 @@
 package backend
 
-import "database/sql"
+import (
+	"database/sql"
+	"errors"
+)
+
+var (
+	errUnknownDefense = errors.New("Unknown defense")
+	errDefenseExists = errors.New("The defense already exists")
+)
 
 
 func Defense(db *sql.DB, id string) (string, error) {
@@ -10,20 +18,18 @@ func Defense(db *sql.DB, id string) (string, error) {
 }
 
 func NewDefense(db *sql.DB, id , cnt string) error {
-	err := SingleUpdate(db, "insert into defenses(id, content) values($1,$2)", id, cnt)
-	return err
+	return SingleUpdate(db, errDefenseExists, "insert into defenses(id, content) values($1,$2)", id, cnt)
 }
 
 func LongDefense(db *sql.DB, id string) (string, error) {
 	var cnt string
 	err := db.QueryRow("select public from defenses where id=$1", id).Scan(&cnt)
+	if err != nil {
+		return "", errUnknownDefense
+	}
 	return cnt, err
 }
 
 func SaveDefense(db *sql.DB, id, shortVersion, longVersion string) error {
-//	log.Printf("%s\n", shortVersion)
-	//log.Printf("%s\n", longVersion)
-	err := SingleUpdate(db, "update defenses set content=$2,public=$3 where id=$1", id, shortVersion, longVersion)
-	return err
-	//return nil
+	return SingleUpdate(db, errUnknownDefense, "update defenses set content=$2,public=$3 where id=$1", id, shortVersion, longVersion)
 }