+++ backend/student.go
@@ -2,8 +2,12 @@ package backend
 
 import (
 	"database/sql"
-	"log"
 	"errors"
+)
+
+var (
+	errUnknownStudent = errors.New("Unknown student")
+	errStudentExists = errors.New("Student already exists")
 )
 
 type Student struct {
@@ -13,35 +17,25 @@ type Student struct {
 }
 
 func NewStudent(db *sql.DB, s Student) error {
-	log.Printf("Creating student %s\n", s)
-	err := NewUser(db, s.P)
+	_, err := NewUser(db, s.P)
 	if err != nil {
 		return err
 	}
-	res, err := db.Exec("insert into students(email, promotion, major) values($1,$2,$3)", s.P.Email, s.Promotion, "?")
-	if err != nil {
-		return err
-	}
-	nb, err := res.RowsAffected()
-	if err != nil {
-		return err
-	}
-	if nb != 1 {
-		return errors.New("Unable to declare the student")
-	}
-	return nil
+	return SingleUpdate(db,
+			errStudentExists,
+			"insert into students(email, promotion, major) values($1,$2,$3)", s.P.Email, s.Promotion, "?");
 }
 
 func GetStudent(db *sql.DB, email string) (Student, error) {
 	var fn, ln, tel, promo, major string
 	err := db.QueryRow("select firstname, lastname, tel, promotion, major from students, users where students.email = users.email and users.email=$1", email).Scan(&fn, &ln, &tel, &promo, &major)
 	if err != nil {
-		return Student{}, err
+		return Student{}, errUnknownStudent
 	}
 	p := User{fn, ln, email, tel, ""}
 	return Student{p, promo, major}, nil
 }
 
 func SetMajor(db *sql.DB, email string, m string) error {
-	return SingleUpdate(db, "update students set major=$2 where email=$1", email, m)
+	return SingleUpdate(db, errUnknownStudent, "update students set major=$2 where email=$1", email, m)
 }